backend/requirements.txt
flask
flask-cors
torch
transformers
numpy
nltk
python-dotenv
requests
